% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virus_circos.R
\name{add_virus_circos_track}
\alias{add_virus_circos_track}
\title{Add a virus track to an existing virus-circos plot object}
\usage{
add_virus_circos_track(x, track)
}
\arguments{
\item{x}{A \code{vi_virus_circos_plot} object.}

\item{track}{A \code{vi_virus_circos_track} object.}
}
\value{
An updated \code{vi_virus_circos_plot} object.
}
\description{
Add a virus track to an existing virus-circos plot object
}
