% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.R
\name{draw_link}
\alias{draw_link}
\title{Draw Link Layer in Genomic Mode}
\usage{
draw_link(
  link_data,
  cfg,
  radius = NULL,
  lwd = 0.35,
  color_map = NULL,
  default_col = "grey"
)
}
\arguments{
\item{link_data}{Data frame converted from a \code{GInteractions} object.}

\item{cfg}{Plotting configuration list.}

\item{radius}{Optional link radius.}

\item{lwd}{Link line width.}

\item{color_map}{Optional named vector of colors for different groups.}

\item{default_col}{Fallback link color.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Draw Link Layer in Genomic Mode
}
