% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virus_circos.R
\name{draw_virus_circos_plot}
\alias{draw_virus_circos_plot}
\title{Draw a virus-circos plot object}
\usage{
draw_virus_circos_plot(x, clear = TRUE)
}
\arguments{
\item{x}{A \code{vi_virus_circos_plot} object.}

\item{clear}{Logical. Whether to clear the current circlize device first.}
}
\value{
The updated \code{vi_virus_circos_plot} object.
}
\description{
Draw a virus-circos plot object
}
