% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virus_circos.R
\name{draw_virus_circos_track}
\alias{draw_virus_circos_track}
\title{Draw a single virus-circos track}
\usage{
draw_virus_circos_track(track, virus_name, virus_length)
}
\arguments{
\item{track}{A \code{vi_virus_circos_track} object.}

\item{virus_name}{Virus sequence name.}

\item{virus_length}{Virus genome length in base pairs.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Draw a single virus-circos track
}
