% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virus_features.R
\name{ncbi_gff3_url}
\alias{ncbi_gff3_url}
\title{Construct the NCBI E-utilities URL for a nucleotide GFF3 annotation}
\usage{
ncbi_gff3_url(accession)
}
\arguments{
\item{accession}{NCBI nucleotide accession.}
}
\value{
A length-one character URL.
}
\description{
Construct the NCBI E-utilities URL for a nucleotide GFF3 annotation
}
\keyword{internal}
