% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virus_circos.R
\name{print.vi_virus_circos_plot}
\alias{print.vi_virus_circos_plot}
\title{Print a virus-circos plot}
\usage{
\method{print}{vi_virus_circos_plot}(x, ...)
}
\arguments{
\item{x}{A \code{vi_virus_circos_plot} object.}

\item{...}{Ignored.}
}
\description{
Print a virus-circos plot
}
