% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virus_circos.R
\name{track_virus_reads}
\alias{track_virus_reads}
\title{Create a virus read track specification}
\usage{
track_virus_reads(height = 0.12, bins = 25, fill = "#A6CEE3")
}
\arguments{
\item{height}{Track height.}

\item{bins}{Number of bins across the virus genome.}

\item{fill}{Histogram fill color.}
}
\value{
A \code{vi_virus_circos_track} object.
}
\description{
Create a virus read track specification
}
